# docs/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "ProofArchiveSender API",
		"version": "1.0.0",
		"description": "REST API for ProofArchiveSender — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
		"license": {
			"name": "SEE LICENSE IN LICENSE",
			"url": "https://github.com/LogicLabs-OU/ProofArchiveSender/blob/main/LICENSE"
		},
		"contact": {
			"name": "ProofArchiveSender",
			"url": "https://ProofArchiveSender.com"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Local development"
		}
	],
	"security": [
		{
			"bearerAuth": []
		},
		{
			"apiKeyAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY",
				"description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
			}
		},
		"responses": {
			"Unauthorized": {
				"description": "Authentication is required or the token is invalid/expired.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Unauthorized"
						}
					}
				}
			},
			"Forbidden": {
				"description": "The authenticated user does not have permission to perform this action.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Forbidden"
						}
					}
				}
			},
			"NotFound": {
				"description": "The requested resource was not found.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Not found"
						}
					}
				}
			},
			"InternalServerError": {
				"description": "An unexpected error occurred on the server.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Internal server error"
						}
					}
				}
			}
		},
		"schemas": {
			"ErrorMessage": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string",
						"description": "Human-readable error description.",
						"example": "An error occurred."
					}
				},
				"required": ["message"]
			},
			"MessageResponse": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string",
						"example": "Operation completed successfully."
					}
				},
				"required": ["message"]
			},
...
			"Role": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "clx1y2z3a0000b4d2"
					},
					"slug": {
						"type": "string",
						"nullable": true,
						"example": "predefined_super_admin"
					},
					"name": {
						"type": "string",
						"example": "Super Admin"
					},
					"policies": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CaslPolicy"
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": ["id", "name", "policies", "createdAt", "updatedAt"]
			},
			"CaslPolicy": {
				"type": "object",
				"description": "An CASL-style permission policy statement. `action` and `subject` can be strings or arrays of strings. `conditions` optionally restricts access to specific resource attributes.",
				"properties": {
					"action": {
						"oneOf": [
							{
								"type": "string",
								"example": "read"
							},
							{
								"type": "array",
								"items": {
									"type": "string"
								},
								"example": ["read", "search"]
							}
						]
					},
					"subject": {
						"oneOf": [
							{
								"type": "string",
								"example": "archive"
							},
							{
								"type": "array",
								"items": {
									"type": "string"
								},
								"example": ["archive", "ingestion"]
							}
						]
					},
					"conditions": {
						"type": "object",
						"description": "Optional attribute-level conditions. Supports `${user.id}` interpolation.",
						"example": {
							"userId": "${user.id}"
						}
					}
				},
				"required": ["action", "subject"]
			},
...
		}
	},
	"paths": {
...
		"/v1/iam/roles": {
			"get": {
				"summary": "List all roles",
				"description": "Returns all IAM roles. If predefined roles do not yet exist, they are created automatically. Requires `read:roles` permission.",
				"operationId": "getRoles",
				"tags": ["IAM"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "List of roles.",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Role"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"post": {
				"summary": "Create a role",
				"description": "Creates a new IAM role with the given name and CASL policies. Requires `manage:all` (Super Admin) permission.",
				"operationId": "createRole",
				"tags": ["IAM"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["name", "policies"],
								"properties": {
									"name": {
										"type": "string",
										"example": "Compliance Officer"
									},
									"policies": {
										"type": "array",
										"items": {
											"$ref": "#/components/schemas/CaslPolicy"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Role created.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					},
					"400": {
						"description": "Missing fields or invalid policy.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/v1/iam/roles/{id}": {
			"get": {
				"summary": "Get a role",
				"description": "Returns a single IAM role by ID. Requires `read:roles` permission.",
				"operationId": "getRoleById",
				"tags": ["IAM"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Role details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"delete": {
				"summary": "Delete a role",
				"description": "Permanently deletes an IAM role. Requires `manage:all` (Super Admin) permission.",
				"operationId": "deleteRole",
				"tags": ["IAM"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Role deleted. No content returned."
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"put": {
				"summary": "Update a role",
				"description": "Updates the name or policies of an IAM role. Requires `manage:all` (Super Admin) permission.",
				"operationId": "updateRole",
				"tags": ["IAM"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"example": "Senior Compliance Officer"
									},
									"policies": {
										"type": "array",
										"items": {
											"$ref": "#/components/schemas/CaslPolicy"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Updated role.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					},
					"400": {
						"description": "No update fields provided or invalid policy.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
...
		"/v1/ingestion-sources/{id}/import": {
			"post": {
				"summary": "Trigger initial import",
				"description": "Enqueues an initial import job for the ingestion source. This imports all historical emails. Requires `create:ingestion` permission.",
				"operationId": "triggerInitialImport",
				"tags": ["Ingestion"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"responses": {
					"202": {
						"description": "Initial import job accepted and queued.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
...
		"/v1/ingestion-sources/{id}/sync": {
			"post": {
				"summary": "Force sync",
				"description": "Triggers an out-of-schedule continuous sync for the ingestion source. Requires `sync:ingestion` permission.",
				"operationId": "triggerForceSync",
				"tags": ["Ingestion"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"responses": {
					"202": {
						"description": "Force sync job accepted and queued.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
...
	},
...
}